requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
oracledb>=2.0.0
python-dotenv==1.0.0
hvac==1.2.1
pydantic==2.4.2
motor==3.3.1
